Добавлен класс Vegetable
Для отработки работы с интерфейсами. Добавил этот класс и наследовался от интерфейса ICollectable. Вывел статус префаба на инфотабло в канвасе.
