+++ SaveDataProject/Assets/Lesson/Controller/ICollectable.cs
