+++ SaveDataProject/Assets/Lesson/Controller/IWorkable.cs
