+++ SaveDataProject/Assets/Lesson/Model.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 280c8011ac6722b409ced8f2de83b2ae
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
