+++ SaveDataProject/Assets/Lesson/Model/Lopata.cs
