+++ SaveDataProject/Assets/Lesson/Model/Player.cs
@@ -1,4 +1,4 @@
-﻿using System.Collections;
+﻿ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
