+++ SaveDataProject/Assets/Lesson/Model/Vegebatle.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 77e47d1655d54c84781ca5a2b946f95a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
