+++ SaveDataProject/Assets/Lesson/View/Inventory.cs
