+++ SaveDataProject/Assets/Lesson/View/Spawner.cs
