+++ SaveDataProject/Assets/LowPolyFarmLite/Prefabs/Cabbage_01.prefab
@@ -13,6 +13,9 @@ GameObject:
   - component: {fileID: 8554979865543785235}
   - component: {fileID: 8378838730965783952}
   - component: {fileID: -6139815317303233744}
+  - component: {fileID: -750131885864627248}
+  - component: {fileID: 4051128424577045151}
+  - component: {fileID: -7963736018311708168}
   m_Layer: 0
   m_Name: Cabbage_01
   m_TagString: Untagged
@@ -111,6 +114,47 @@ MonoBehaviour:
   id: 4
   countItem: 0
   isStakable: 1
-  descriptionItem: "\u041A\u0430\u043F\u0443\u0441\u0442\u0430 \u043A\u0440\u0430\u0441\u043D\u043E\u043E\u0434\u0430\u0440\u0441\u043A\u0430\u044F"
+  descriptionItem: "\u041A\u0430\u043F\u0443\u0441\u0442\u0430 \u043A\u0440\u0430\u0441\u043D\u043E\u0434\u0430\u0440\u0441\u043A\u0430\u044F"
   pathIcon: Assets/Lesson/Images/Cabbage
   pathPrefab: Assets/LowPolyFarmLite/Prefabs/Cabbage_01
+--- !u!114 &-750131885864627248
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 8554979865545778643}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 77e47d1655d54c84781ca5a2b946f95a, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!135 &4051128424577045151
+SphereCollider:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 8554979865545778643}
+  m_Material: {fileID: 0}
+  m_IsTrigger: 1
+  m_Enabled: 1
+  serializedVersion: 2
+  m_Radius: 0.2565651
+  m_Center: {x: -0.0114277825, y: 0.16427515, z: 0.021046095}
+--- !u!54 &-7963736018311708168
+Rigidbody:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 8554979865545778643}
+  serializedVersion: 2
+  m_Mass: 1
+  m_Drag: 0
+  m_AngularDrag: 0.05
+  m_UseGravity: 0
+  m_IsKinematic: 0
+  m_Interpolate: 0
+  m_Constraints: 0
+  m_CollisionDetection: 0
